Refactor project structure and navigation setup
Moved assets, components, and screens into a new 'src' directory for better organization. Updated import paths and app.json asset references accordingly. Split navigation logic into 'RootStack.js' and 'Routes.js' for improved maintainability.
